// src/tests/ticket/editTicket.generated.ts
/* eslint-disable */
/* @ts-nocheck */
/* prettier-ignore */
/* This file is automatically generated using `npm run graphql:types` */
import type * as Types from '../../generated/types';

import type { JsonObject } from "type-fest";
import gql from 'graphql-tag';
export type EditTicketMutationVariables = Types.Exact<{
  input: Types.TicketEditInput;
}>;


export type EditTicketMutation = { __typename?: 'Mutation', editTicket: { __typename?: 'Ticket', id: string, name: string, description: string | null, status: Types.TicketTemplateStatus, visibility: Types.TicketTemplateVisibility, startDateTime: string, endDateTime: string | null, requiresApproval: boolean | null, quantity: number | null, price: number | null, currencyId: string | null, eventId: string } };


export const EditTicket = gql`
    mutation EditTicket($input: TicketEditInput!) {
  editTicket(input: $input) {
    id
    name
    description
    status
    visibility
    startDateTime
    endDateTime
    requiresApproval
    quantity
    price
    currencyId
    eventId
  }
}
    `;